app.py
import os
import logging
from telegram.ext import Application, MessageHandler, filters
from flask import Flask, request

# កំណត់ Logging
logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)

# អាន​ទិន្នន័យ​សម្ងាត់​ពី Environment Variables
BOT_TOKEN = os.environ.get("BOT_TOKEN")
# សម្រាប់ Webhook ត្រូវ​ប្រើ URL របស់ Render
WEBHOOK_URL = f"https://YOUR_RENDER_SERVICE_NAME.onrender.com/{BOT_TOKEN}"
# ត្រូវ​កំណត់​លេខ​សម្ងាត់​ដើម្បី​សុវត្ថិភាព
SECRET = os.environ.get("WEBHOOK_SECRET")

# កំណត់ Port តាម​ដែល Render បាន​ផ្តល់​ឲ្យ
PORT = int(os.environ.get("PORT", 5000))

app = Flask(__name__)
application = Application.builder().token(BOT_TOKEN).build()

# មុខងារ​ទាញយក​វីដេអូ​របស់​អ្នក​នៅ​ទីនេះ (ដូច​បាន​រៀបរាប់​មុន)
def download_video(update, context):
    # កូដ​ទាញយក​វីដេអូ​របស់​អ្នក​ដាក់​នៅ​ទីនេះ
    update.message.reply_text("✅ Bot បាន​ទទួល URL ហើយ​កំពុង​ទាញយក...")
    # ... បន្ត​កូដ yt-dlp និង send_video ...
    pass # ត្រូវ​ជំនួស​ដោយ​កូដ​ពិត

application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, download_video))


# កំណត់ Webhook Endpoint សម្រាប់​ទទួល​សារ​ពី Telegram
@app.route(f"/{BOT_TOKEN}", methods=["POST"])
def webhook_handler():
    # ត្រូវ​ពិនិត្យ​មើល​ Secret Token ដើម្បី​សុវត្ថិភាព
    if request.headers.get("X-Telegram-Bot-Api-Secret-Token") != SECRET:
        return "Invalid Secret Token", 403
    
    # បញ្ជូន​ទិន្នន័យ​ដែល​ទទួល​បាន​ទៅ​កាន់ Telegram Application
    update = request.get_json(force=True)
    dispatcher = application.create_dispatcher()
    dispatcher.process_update(update)
    return "ok"

# Route សាមញ្ញ​សម្រាប់​ត្រួតពិនិត្យ Health Check
@app.route("/")
def index():
    return "Bot is running!"

# ចាប់ផ្តើម​កំណត់ Webhook ពេល Server ចាប់ផ្តើម​ដំបូង
if __name__ == "__main__":
    # កំណត់ Webhook ទៅ​កាន់ Telegram
    application.bot.set_webhook(
        url=WEBHOOK_URL,
        secret_token=SECRET
    )
    # ដំណើរការ Flask App ដោយ​ប្រើ Gunicorn នៅ​ពេល​ដាក់​ពង្រាយ
    app.run(host="0.0.0.0", port=PORT)