requirements.txt
python-telegram-bot
yt-dlp
Flask         # បន្ថែម Flask សម្រាប់ Webhook
gunicorn      # សម្រាប់​ដំណើរការ Flask app លើ Render